py_grim/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
prometheus-client==0.19.0
redis==5.0.1
slowapi==0.1.9
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6 
PyYAML>=6.0.0
aiohttp>=3.12.0
bcrypt>=4.0.0

# TuskLang Python SDK - Official PyPI Package
tusktsk>=2.0.3

# Optional database dependencies
psycopg2-binary>=2.9.0
pymongo>=4.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0 